AI_assigment/requirements.txt
gradio
PyMuPDF
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
numpy
